fix(structure) : BTAP-1 refactor package
