--- src/main/java/id/bfi/co/id/App.java
@@ -1,12 +1,0 @@
-package id.bfi.co.id;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class App {
-
-  public static void main(String[] args) {
-    SpringApplication.run(App.class, args);
-  }
-}
