--- src/main/java/id/bfi/co/id/constant/AppConstants.java
@@ -1,7 +1,0 @@
-package id.bfi.co.id.constant;
-
-public class AppConstants {
-  private AppConstants() {}
-
-  public static final String HELLO_MESSAGE = "Hai";
-}
