--- src/main/java/id/bfi/co/id/dto/DummyResponse.java
@@ -1,18 +1,0 @@
-package id.bfi.co.id.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class DummyResponse {
-  private Long id;
-  private String code;
-  private String name;
-  private String category;
-  private String description;
-}
