--- src/main/java/id/bfi/co/id/exception/BusinessException.java
@@ -1,5 +1,0 @@
-package id.bfi.co.id.exception;
-
-public class BusinessException extends Exception {
-  private static final long serialVersionUID = 1L;
-}
