--- src/main/java/id/bfi/co/id/service/GroupCamundaService.java
@@ -1,10 +1,0 @@
-package id.bfi.co.id.service;
-
-import java.util.List;
-
-import org.camunda.bpm.engine.delegate.DelegateExecution;
-
-public interface GroupCamundaService {
-  public Integer countGroupByUserName(DelegateExecution execution, String userName);
-  public List<String> getListMemberGroupByGroupName(DelegateExecution execution, String groupName);
-}
