--- src/main/java/id/bfi/co/id/service/UserCamundaService.java
@@ -1,8 +1,0 @@
-package id.bfi.co.id.service;
-
-import org.camunda.bpm.engine.delegate.DelegateExecution;
-
-public interface UserCamundaService {
-  public Integer countUserByGroupName(DelegateExecution execution, String groupName);
-  public String getUserByRoundRobin(DelegateExecution execution, String groupName);
-}
