+++ src/main/java/id/co/bfi/bfitaskallocation/App.java
@@ -1,0 +1,12 @@
+package id.co.bfi.bfitaskallocation;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class App {
+
+  public static void main(String[] args) {
+    SpringApplication.run(App.class, args);
+  }
+}
