+++ src/main/java/id/co/bfi/bfitaskallocation/auth/SecurityContextService.java
@@ -1,0 +1,17 @@
+package id.co.bfi.bfitaskallocation.auth;
+
+import java.util.UUID;
+
+import org.springframework.security.authentication.AbstractAuthenticationToken;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+
+public interface SecurityContextService {
+  public UsernamePasswordAuthenticationToken decodeUserToken(String token);
+  public AbstractAuthenticationToken decodeServiceApiSecret(String key);
+  public long getCurrentUserId();
+  public UUID getCurrentUserTokenId();
+  public String getCurrentUserPhoneNumber();
+  public String getCurrentToken();
+  public String getDeviceId();
+  public String getCurrentUserEmail();
+}
