+++ src/main/java/id/co/bfi/bfitaskallocation/constant/AppConstants.java
@@ -1,0 +1,7 @@
+package id.co.bfi.bfitaskallocation.constant;
+
+public class AppConstants {
+  private AppConstants() {}
+
+  public static final String HELLO_MESSAGE = "Hai";
+}
