+++ src/main/java/id/co/bfi/bfitaskallocation/constant/LoggingConstants.java
@@ -1,0 +1,7 @@
+package id.co.bfi.bfitaskallocation.constant;
+
+public abstract class LoggingConstants {
+  public static final int CORRELATION_ID_SCOPE = 99;
+  public static final String CORRELATION_ID_LOG_VAR_NAME = "correlationId";
+  public static final String CORRELATION_ID_HEADER_NAME = "X-Correlation-Id";
+}
