+++ src/main/java/id/co/bfi/bfitaskallocation/dto/CreateDummyRequest.java
@@ -1,0 +1,28 @@
+package id.co.bfi.bfitaskallocation.dto;
+
+import javax.validation.constraints.NotBlank;
+import javax.validation.constraints.NotNull;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class CreateDummyRequest {
+
+  @NotBlank
+  private String code;
+
+  @NotBlank
+  private String name;
+
+  @NotNull
+  private String category;
+
+  @NotNull
+  private String description;
+}
