+++ src/main/java/id/co/bfi/bfitaskallocation/dto/DummyResponse.java
@@ -1,0 +1,18 @@
+package id.co.bfi.bfitaskallocation.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class DummyResponse {
+  private Long id;
+  private String code;
+  private String name;
+  private String category;
+  private String description;
+}
