+++ src/main/java/id/co/bfi/bfitaskallocation/exception/BusinessException.java
@@ -1,0 +1,5 @@
+package id.co.bfi.bfitaskallocation.exception;
+
+public class BusinessException extends Exception {
+  private static final long serialVersionUID = 1L;
+}
