--- src/test/java/id/bfi/co/id/AppTest.java
@@ -1,16 +1,0 @@
-package id.bfi.co.id;
-
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-public class AppTest {
-
-  /**
-   * Rigorous Test :-)
-   */
-  @Test
-  public void shouldAnswerWithTrue() {
-    assertTrue(true);
-  }
-}
