+++ src/test/java/id/co/bfi/bfitaskallocation/AppTest.java
@@ -1,0 +1,16 @@
+package id.co.bfi.bfitaskallocation;
+
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+public class AppTest {
+
+  /**
+   * Rigorous Test :-)
+   */
+  @Test
+  public void shouldAnswerWithTrue() {
+    assertTrue(true);
+  }
+}
