+++ src/test/java/id/co/bfi/bfitaskallocation/controller/DummyControllerTests.java
@@ -1,0 +1,48 @@
+package id.co.bfi.bfitaskallocation.controller;
+
+import static org.mockito.Mockito.when;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.http.HttpStatus;
+import org.springframework.test.context.ActiveProfiles;
+import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.setup.MockMvcBuilders;
+import org.springframework.web.context.WebApplicationContext;
+import org.springframework.web.server.ResponseStatusException;
+
+import id.co.bfi.bfitaskallocation.service.DummyService;
+
+@SpringBootTest
+@AutoConfigureMockMvc
+@ActiveProfiles(profiles = "unit-test")
+public class DummyControllerTests {
+
+  @Autowired
+  private MockMvc mockMvc;
+
+  @Autowired
+  private WebApplicationContext webApplicationContext;
+
+  @MockBean
+  private DummyService mockDummyService;
+
+  @BeforeEach
+  public void setup() {
+    mockMvc = MockMvcBuilders.webAppContextSetup(webApplicationContext).build();
+  }
+
+  @Test
+  public void getLatestDummy() throws Exception {
+    when(mockDummyService.getLatestDummy())
+      .thenThrow(new ResponseStatusException(HttpStatus.NOT_FOUND));
+
+    mockMvc.perform(get("/dummy")).andExpect(status().isNotFound());
+  }
+}
